Added composite scoring algorithm.
Adds an implementation of a composite scoring algorithm, along with some
documentation about why we might want to do this, and some suggestions on
how to proceed.  Also includes a link to some examples from 2017 that
would be used when trying to identify something more sophisticated than
"the largest yellow blob that looks like a ball".