Adding an (untested) command to run the conveyor during auto mode.
This should (in theory) allow for starting with either:
* a fully-empty chamber (in which case it should bring a
  ball in and send it through, then stop when the next one is loaded)
* a ball already loaded (in which case it'll push that through and then
  stop when the next one is loaded)

There's also support for a timeout/deadline, at which point the task will
assume that we're not going to see another ball, and just end.

This does assume that the ball sensor in the conveyor path is working, and
is being read correctly by the code.  

There's also corner cases that still need to be dealt with, though there
are some suggestions in the code for possibilities on at least one.
* The "one ball pushing up against the next prevents the sensor from
  seeing a gap" case is one example, which can't really be fixed *in
  code*.
* The "one ball follows *just* behind the other, but with a micro-gap big
  enough to be noticed, but not enough to get the first ball to the
  shooter" case is another.  (Which *might* be something that can be handled
  in code.)