Camera simulation updates, and initial TurnToTarget use.
This updates the original SimVisionWrapper class so that it no longer serves as a direct wrapper around the "real" vision subsystem.  Instead, the CameraSimulator is now a distinct/separate injection mechanism.

Among other things, this means that the CameraSimulator doesn't need to implement either IVision or IBetterVision, which makes it easier to swap things around (and make run-time choices based on the configured setup).