Breaking advanced IDrivebase stuff into IBetterDrivebase.
The idea here is that this demonstrates how functionality can be implemented in stages.  It might also make it possible for me to provide a simple implementation of the basics (as a teaching tool), and then layer on more complex functionality in a derived class (or at least in an alternate one).